(refactor) Refactor modal files to match naming conventions
