+++ packages/apps/esm-devtools-app/src/index.ts
@@ -10,6 +10,6 @@ const options = {
 export const devtools = getAsyncLifecycle(() => import('./devtools/devtools.component'), options);
 
 export const importmapOverrideModal = getAsyncLifecycle(
-  () => import('./devtools/import-map-list/modal.component'),
+  () => import('./devtools/import-map-list/import-map.modal'),
   options,
 );
